refactor: centralise subpage header
