+++ packages/shared/src/components/Layout/SubpageHeader/index.tsx
@@ -1,0 +1,1 @@
+export { SubpageHeader } from './SubpageHeader'
