+++ packages/shared/src/components/Layout/index.ts
@@ -1,2 +1,3 @@
 export * from './Header'
 export { Layout } from './Layout'
+export { SubpageHeader } from './SubpageHeader'
